#782: move rollbar/deploy/capistrano to rollbar/deploy; add support for deploy status failed
