+++ lib/rollbar/capistrano.rb
@@ -1,6 +1,6 @@
 require 'capistrano'
-require 'rollbar/deploy/capistrano'
+require 'rollbar/deploy'
 
 if Capistrano::Configuration.instance
-  Rollbar::Capistrano.load_into(Capistrano::Configuration.instance)
+  Rollbar::Deploy::Capistrano.load_into(Capistrano::Configuration.instance)
 end
